Display Artwork Index Page
Added functionality to loop through and display all of the artwork and tested it by entering a test piece into the database.
I also updated the Artwork model to take a foreign key from an Artist model so as not to repeat an artist name (in case an artist has multiple artworks).