Added Comments and 404 Messages
Cleaned up code and added comments.

Added additional 404 messages if the category in the url does not match a category in the database, or if the floor in the url does not match a floor in the current building.

If we are going this far, we should think about if we want to display an error message if a selected category is not in the building or floor. Or should we keep the regular view and just display a message that reads something like "No artwork of this category in this floor/building combination".